dlls: Flush CompositionString after X11DRV_XIMLookupChars.
There is no flushing after X11DRV_XIMLookupChars on wine11xdrv. for
this reason, the characters of the GUI are not updated.

In addition, in IME_SetResultString on wine11xdrv, the bInComposition
flag is not deactivated after generating WM_IME_ENDCOMPOSITION, so that
the WM_IME_STARTCOMPOSITION message is not generated in the next
composition.

Also, when WM_IME_COMPOSITION is processed in editor_handle_message on
riched20, the existing selection is removed and the characters being
entered disappear.
WM_IME_STARTCOMPOSITION occurs when attempting to input after full
selection, and ME_DeleteSelection is called, which causes the exact
location of the cursor to be lost.

Signed-off-by: Alex Kwak <take-me-home@kakao.com>
---
v2: Fix comment style