winex11.drv: Use composition string from XIM on IME_SetResultString()
IME_SetResultString() function overwrite composition string as empty.
so, next composition may not appear on the displayed. but, still have
an xim. Therefore, there was a string that seemed to have disappeared
in the middle of the input.

Therefore, change it as follows. When calling IME_SetResultString(), if
in a Composition state, no longer generate WM_IME_ENDCOMPOSITION and
Use CompositionString in XIM to match the current XIM status rather
than overwriting Composition.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=52700
Signed-off-by: Alex Kwak <take-me-home@kakao.com>